PR12.3: eliminate package-import runtime mutation
