+++ tools/engineering_quality_gate.py
@@ -24,6 +24,9 @@ LEGACY_RELOCATION_SOURCES = {
     "src/chatgpt_web_adapter/browser_owned_product_transport_core.py": (
         "src/chatgpt_web_adapter/browser_owned_product_transport.py"
     ),
+    "src/chatgpt_web_adapter/product_runtime_core.py": (
+        "src/chatgpt_web_adapter/product_runtime.py"
+    ),
 }
